requirements.txt
# pythonLibs Requirements
requests>= 2.24.0
PyYaml>=5.3.1
pytz>=2020.1
pytz-convert>=0.2.9
user-agents>=2.2.0
numpy>=1.19.2
pandas>=1.1.3
sklearn~=0.0
stringcase>=1.2.0
IPy>= 1.0
colorlog>=4.4.0
beautifulsoup4>= 4.9.3
xlsxwriter>=1.3.7
geoip2>=4.1.0
regex==2015.10.29
pysocks
urllib3~=1.25.11
scikit-learn~=0.23.2
setuptools~=50.3.2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
selenium==3.141.0
elasticsearch
sqlalchemy
iso8601
xmltodict